[FIX] sale_timesheet_enterprise: show correct allocated hours for portal users
When a task with an allocated time > 0.0 but no timesheets is created in a shared project (with edit rights), the portal user incorrectly sees 0.0 as the allocated time.

Steps to reproduce:

- Ensure the sale_timesheet module is installed.
- Create a new project.
- Create a task with allocated time and no timesheets.
- Share the project with a portal user (edit permission).

In the portal user's kanban view, the allocated time of the task is displayed as 0.0 instead of the correct allocated time.

opw-4582705

closes odoo/enterprise#85385

X-original-commit: 57a41f58210536e2314de027950beff288750012
Signed-off-by: Soukaina Tyes (soty) <soty@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>