[FIX] sale_stock_renting:  account for rents in forecast availability
Steps to reproduce:

- In the settings enable "Rental transfers"
- Create a storable product that can be rented put 100 units in stock.
- Create a rental order for 10 units of that product planed for the
  period : [today + 2 days, today + 3 days] and confirm it.
- Create a rental order for 10 units of that product planed for the
  period : [today + 1 days, today + 2 days] and confirm it.
- Look at the forecast rentable quantity.
> It should be 90 but it is 100.

Cause of the issue:

Since c18a7d2dc36d33134c5bdb1569865bd554e04130, the rentable forecast
for future dates ignores the `rented_qty_during_period` if the setting
"Rental transfers" is enabled. This happens because that rented
quantity during period is supposed to be absorbed by the stock forecast:
https://github.com/odoo/enterprise/blob/8ea5f4f9a0a907c57844c6274688171193f6a904/sale_stock_renting/models/sale_order_line.py#L132-L134
However, the stock forecast of the 'virtual_available' relies solely
on deliveries and receipt happening prior to the start of the renting
period. Therefore the rental orders that are planned to start during
the renting period should still contribute to the
`rented_qty_during_period` in that use case.

opw-4552760

closes odoo/enterprise#83708

X-original-commit: 3de00aab92a90fa61c7e5865e47ebe1382d37549
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>