[PERF] account_intrastat: send the filter domain instead of list
Previously, the intrastat_code_id was filtered by checking
against a large list that is sent to the frontend. This caused
huge memory usage due to the list being used in many objects
and the list itself being large, around 10k items. This memory
leak caused slowness in the frontend while going through the
product form views.

Currently, the list is not being sent. The domain itself is
evaluated from the frontend without sending a list. This reduces
memory footprint and slowness.

closes odoo/enterprise#83480

X-original-commit: 4c190c09f8224b11563167cb974bc21497b2e05e
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>