[FIX] appointment: use default recipients on canceled template
Not sure why there are no specific recipients given, which means only
followers are notified by default. It is not necessarily an issue as
most attendees are also followers, but in some case some notifications
are not sent (test_appointment_mail_notification_from_portal).

Use default recipients, which enables sending emails to attendees that
are not followers.

Task-4655022

X-original-commit: f6611d75494f65e64170f551f524e03fd58f3967
Part-of: odoo/enterprise#82293
Related: odoo/odoo#203670
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>