[FIX] account_reports: if_below/if_above
Take the use case of three expressions:

A: simple expression with a result of 2500
B: simple expression with a result of 0
C: aggregation for which we want to get the
value of A only if B is below 50

This will not work, because _aggregation_apply_bounds
will return the unbound_value and cast it to a bool then
int to get the multiplier.
If the unbound value is 0, it will set the result to 0 even
though 0 is indeed below 50.

We now return "None" from _aggregation_apply_bounds
when the value is out of bound
instead of 0, to remove any possible
confusion.

closes odoo/enterprise#82284

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>