[FIX] hr_payroll_expense: fix test failure in non-demo mode
In `test_no_expense_rule_means_no_linkage` we create a new salary
structure to test the case where a missing rule shouldn't link expenses
to the payslip.

In the copy of the base setup structure we didn't mention a journal,
this lead to the use of another journal that worked because demo data
would ensure it was properly setup.

Without demo data, this test would fail.

This commit ensures the journal used in the test is the one
from the base setup.

closes odoo/enterprise#82300

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>