[IMP] documents_hr(_payroll): remove send HR Access link and set payroll documents access link permanent
Payroll documents are set to anyone with the link, making the link that
employees receive in the payslip/declaration mails permanent.
They will always be able to access their payroll document using that link.
Apparently, no need to send them the documents aggregated into a single place
where it's easy to download all the documents at the same time anymore.

In the worst case scenario, the employee can ask HR to re-send the document
link one by one if, for example, the employee deleted those HR documents mails.
Note that HR Payroll officers don't have access anymore to the employee's
payslip via document app. So they will need to use payroll app to re-send the
payslips. That is the purpose of the following commit. (That should probably be
squashed with this one)

Task-4309991

Part-of: odoo/enterprise#80399
Related: odoo/odoo#203729
Related: odoo/upgrade#7532
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>