[FIX] documents(,_spreadsheet): fix downloading folders
Somehow we lost the frontend interface to the backend-ready
downloading of folder.

Also fix testing helpers to make the details panel usable
without crashing, and to make non-root folders appear
as children of their parents in the search panel.
While we're here, linting test code introduced in this version
allows to be ready to limit diffs in future fixes & FW ports.

Task-4677814

closes odoo/enterprise#82507

X-original-commit: e2746f8badd2396a881e5418a96e7faa46474365
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>