[ADD] web_studio_ai_fields: support AI Fields in Studio
Purpose:
--------
This module adds integration between web_studio and ai_fields, allowing
users to create AI Fields directly from Studio.

A new field "AI field" is added in the editor sidebar. When dragging it into
the view, the user is prompted to select the field type and the prompt (and
the selections if the field is a selection, or the relation if the field is
a M2x).
The AI field is then created and added in the view. The field will be computed
by the ai_field cron for the 50 last records of the model.

Task-4526274

Part-of: odoo/enterprise#83027
Related: odoo/odoo#205139
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>