[FIX] account_reports: force current company in Customer Statement report options
Steps to reproduce:
- Accounting > Reporting > Partner Ledger
- Select multiple partners
- Select "Customer Statements" as report
- Click "Send"
- Go to the one of the partner in question an dopen the generated PDF
- All current companies show in the header and other currencies are present

Cause:
When sending to multiple partners, the action is dispatched with cron so a new `env` is created. This `env` has a key `companies` containing all companies that exist in the DB.
When the `_init_options_companies` is called for the report it adds `self.env.companies` to the option ([code](https://github.com/odoo/enterprise/blob/4ef74fd9527a28616bf81c9716d6000aa25b839e/account_reports/models/account_report.py#L1276)).
So in the end all companies appear on the report.

Solution:
Add 'forced_companies' in the options before calling the `cron`.

The format of 'forced_companies' had to be change to only contain ids because fields cannot be added in `send_and_print_values` because it's a JSON fields.

opw-4635283

closes odoo/enterprise#83989

X-original-commit: b210d787dd48255b3c85d779a86511a24ddb2aa8
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>