[FIX] account: remove rd-sm-learn group from account
Removes the explicit groups:
 - `group_survey_user`
 - `group_website_slides_manager`
 - `group_documents_manager`
from account common class and adapt tests to work.

The groups are added explicitly in tests that require them, not using
the `get_default_groups` method as those groups are not, and _should
not_, be required by parent `setUpClass`es.

Part-of: odoo/enterprise#82534
Related: odoo/odoo#204122
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>