[FIX] delivery_ups: display VAT in UPS invoice
**Problem:**
When making a shipment with UPS, validating the
delivery and thus generating the UPS invoice,
the VAT of the companies won't be shown in their
given field.

**Steps to reproduce:**
- Create a quotation, chose a customer that is
not is the same country as your company.
- Add a delivery and chose UPS.
- Validate the delivery.
- Check the invoice generated in the chatter.
- The VAT is not in the invoice

**Cause of the issue:**
The VAT of the company and the VAT of the customer
were never given to UPS.

**Fix:**
The VATs are now added to the UPSRequest class
before making the call to the API.

As the TaxIdentificationNumber field is depricated,
we have to use the new GlobalTaxInformation
container, which allows us to specify the tax
information about the sender and the shipper.

opw-4591744

closes odoo/enterprise#97807

X-original-commit: 13b4021b123e3b93097eeb2bdf6816ecd2af5f25
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>