[FIX] l10n_mn_reports: Other gain (loss) formula
Fix an issue with the formula of the Other gain (loss)
line.
It was "-57 + -59" which in theory is a correct
formula, but in practice it is not handled well
by the report engine and the + will match many
other accounts that we don't want.

It is corrected to be "-57 - 59".

task-4677500

closes odoo/enterprise#82634

X-original-commit: 82ed1b59fdf5e952f1dcb9756d1bb311d1345803
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>