[FIX] sale_commission: avoid loading all related records
Before this commit, achievement were performed using records on
account_move, sale_order and sale_order_log tables (+ move lines, order
lines, products etc).

Reading those table is extermely costly. We tried to limit what we
fetched using context depending on the current action:

- https://github.com/odoo/enterprise/pull/77483/commits/7c8621e71b1e8267cd6e0e7968cbb46705568e32
- https://github.com/odoo/enterprise/pull/80043

And we had to revert it here:
https://github.com/odoo/enterprise/pull/80978/commits/c6ee48faa967342c958b4410e71b49bf453cd316

Using a period_domain in the context computed on fly in _where_calc was
a bad idea. The record were correctly fetched but the following flow was
broken:

Go to an approved commission plan --> click on the commissions stat
button --> everything is correct
Now click on details for a commission line with non empty achived amount

All the achievements are correct but when we click on "Details" of
sale.commission.achievement.report (id=42) account_move (142), we may
end up on another record.

Because when the ORM does the fetch query, it will fetch
sale.commission.achievement.report where id=42 and at that moment, the
context is lost when we click on the line.

Therefore as ids are recomputed, the line 42 is not the one
corresponding to the account move 142 anymore.

This commit ensure we avoid doing a cross join on account_move or
sale_order or sale_order_log because these are really bad for
performance and we select the record in a CTE.
- We select them based on current users, teams if defined
- We avoid fetching records outside of plan date values

X-original-commit: f05b88a39dcdbb7432504fb5f2bc2e0d99d1ab2a
Part-of: odoo/enterprise#84435
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>