[IMP] sign/saas~18.4: clean up the sign item roles
Since the big UI refactor in the sign app, the sign item roles are not used anymore
and new roles are added on the template on the fly.
This commit aims to clean up the data file of these roles and update the tests not to use them.

closes odoo/enterprise#82445

Task-id: 4646742
Related: odoo/upgrade#7472
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>