[FIX] sale_subscription: load translations on website
The `sale_subscription` module defines a template used on the website,
but it was not marked as being a frontend module. Therefore the
translations were not loaded on the website and the template was never
translated.

In this commit we add this module to the frontend modules, so its
templates displayed on the website are properly translated.

closes odoo/enterprise#83351

X-original-commit: e54246957f9b6aec780c32226da580168d3d2fa7
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>