[IMP] l10n_it_{reports|xml_export}: Add XML export option for monthly tax report
Description of the issue this commit addresses:

The Italian localization currently doesn't allow users to export the periodic
tax report to xml format. This feature is missing.

Desired behavior after the commit is merged:

The Italian monthly tax report has a button that allows the user to export the
report to an xml format.

With the new LIPE export, posting a closing entry opens an export wizard which
breaks some tests. This PR implements a workaround to not open that wizard.

Community PR: https://github.com/odoo/odoo/pull/195948
task-4507942

closes odoo/enterprise#82504

X-original-commit: 03c118add6b8247e1dd69dba5f417a4dc3e09188
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>