[FIX] project_enterprise: make sure task created take whole gantt period
Before this commit, the test "/project_enterprise/tests/test_ui.py:ProjectEnterpriseTestUi.test_01_ui"
fails in no demo because the task created before starting the tour is
not planned for the whole gantt period displayed. Which means the
user does not overwork as expected when the tour plans another task.

This commit makes sure the task is planned during the whole period
displayed in the gantt view to correctly check the gantt progress bar
detects the user overworks.

runbot-error-161930

closes odoo/enterprise#82511

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>