[IMP] l10n_be_hr_payroll: improve spouse revenue field explanation
In BE, the spouse revenue selection field has been improved. The order
of the fields in the selection has been changed to be more relevant.
Also, the explanation text of the different selection value has been
improved and corrected.

This has also been corrected in the salary configurator. Due to the
change of the default value for the spouse fiscal status, the contract
salary tour had to be adapted.

task-4680475

closes odoo/enterprise#82522

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>