[IMP] documents_hr: update documents error message
In order to use the Documents app from Employee, you need a valid associated
contact. However, the error message when no contact exists was outdated and
showed an irrelevant message.

This commit updates the error message to indicate that a contact is needed
to use the Documents app from the Employee.

task-4684139

closes odoo/enterprise#82525

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>