[FIX] planning: compute end_datetime with template and timezone
Before this commit, the `Europe/Brussels` timezone was applied 2 times
on the calculation of the end_datetime when the slot is linked to a
template and so the end date time of the slot could suddenly be before
the start date time, which should not be possible. The timezone is
applied 2 times, because the start date time computed already got the
right timezone, so it is not needed to apply again the timezone since
the timezone is already defined thanks to the start date time used in
the computation.

This commit makes sure the timezone is applied once during the
calculation of the end datetime, to be sure the difference between UTC
and the timezone to apply is not applied 2 times on the end date time.

runbot-error-162055

closes odoo/enterprise#82755

X-original-commit: 1aa8381ddd9425d0a4f00f3b46d3f57578faca32
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>