[IMP] timesheet_grid: don't make some fields readonly when timer running
Before this commit, when a timer is running, the timesheet linked to
that timer is completely readonly in the list view except in the timer
header. The problem is the record cannot be saved since the fields are
not supposed to be editable for the list model.

This commit avoid making the fields available in the timer header
readonly when the timer is running for that timesheet. By doing that,
the timesheet can be editable in the timer header and/or in the list
row without any issues.

task-4284216

Part-of: odoo/enterprise#81686
Related: odoo/odoo#202313
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>