[IMP] account_accountant: enable changing filters behind rainbowman
This is more an UX problem rather than a bug.

In the reconciliation tool, when you reconciled all the transactions
of the current filter, rainbow mans appears.

Then, all the background behind the rainbow overlay is disabled,
you can no longer edit the filters in the search for instance,
and the only buttons available become "All transactions" and "Back to
dashboard". Both buttons will make you loose your search filters:
- You loose the "Not Matched" filter which is set automatically
when you hit the "x to reconcile" button in the accounting dashboard,
and hence you see the already reconciled transactions,
while you were actually interested only in the non-reconciled
transactions
- If you made a complicated filter, you loose them all and you
are good to redo them all.

This revision offers the possibility to have the background
behind the rainbow overlay clickable, so you would be able
to change the filters, for instance only the last filter,
in order to continue your reconciliation where you were,
instead of being completely reset.

closes odoo/enterprise#82922

X-original-commit: c6b32348741b2b8b8242204b438ceb9cb473c856
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>