[FIX] mrp_plm: compute ECO count separately for Product and BoM
Steps to reproduce the bug:

- Create a storable product “P1”:
  - Create a BoM
  - Go back to the product form and create an ECO:
    - Apply on: “Bill of Materials”
    - The created BoM will be selected automatically.
    - Save.

- Return to the product form.
- The ECO count in the smart button is 1.
- Click on it.

Problem:
No ECO is displayed because the domain filters only ECOs applied to the
product:
https://github.com/odoo/enterprise/blob/ec6b4717f49ccd5b651be80aed21494ed500f580/mrp_plm/models/product.py#L17-L20

Solution:
only ECOs of type "Product" should be counted in the product form.

Issue 2:

- Try updating the "Apply on" field of the ECO to “Product.”

Problem:
The "Bill of Materials" field becomes invisible but is not cleared.
As a result, when counting the number of ECOs linked to the BoM, the
count will be incorrect.

https://github.com/odoo/enterprise/blob/ec6b4717f49ccd5b651be80aed21494ed500f580/mrp_plm/models/mrp_bom.py#L24-L26

opw-4653598

closes odoo/enterprise#82856

X-original-commit: 96dcbb1f48fa13b5ce865281e31b81ecb90dc4e3
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>