[FIX] stock_barcode: remove many2one_uom widget
Issue
=====
When displaying the Barcode move line form view on mobile, a traceback
can happen.

How to reproduce
================
- Open the Barcode app on a mobile device (or use devtools to simulate a
  mobile device;)
- Create a new receipt;
- Click on "Add Product" and select a product with an existing quant;
- Confirm then re-open the move line's form view -> Traceback.

Cause of the issue
==================
In mobile view, the product's quants displayed in the move line form
view are displayed as kanban record. This kanban view template use the
`many2one_uom` widget but doesn't include the `product_id` field which
make this widget not usable. By the way, those records are not editable
so it makes no sense to use this widget in this case.
To fix the issue, remove the widget where it is not needed/not usable.

closes odoo/enterprise#82839

X-original-commit: 6b105fac99a79a21710bc476889804857113a7a9
Related: odoo/odoo#204699
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>