[FIX] spreadsheet_edition: allow scrolling in readonly pivot sidepanel
The `pointer-events: none;` property prevents elements (and their children)
from receiving any pointer events, including scroll gestures. As a result,
scrolling the pivot dimensions in readonly mode was not possible.

To fix this, we removed `pointer-events: none;` from the main container
and applied it only to the child elements of the pivot dimensions.

Task: 4655830
Part-of: odoo/enterprise#83447
Related: odoo/odoo#206102
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>