[FIX] sale_subscription_timesheet: update delivered qty of prepaid timesheet
**Issue**
In a subscription, the qty delivered of a SOL with "Prepaid" invoicing policy
wasn't updated when timesheets were added.

**Steps to reproduce**
1. Have a service subscription product with an invoicing policy "Prepaid/Fixed
price" and creating a task on order.
2. Create a monthly subscription with a start date 15 days in the past
3. Add some timesheet hours on the task.
4. Run the "Subscription: generate recurring invoices" cron.
- Actual: quantity delivered on the SOL stays at 0.
- Expected: timesheet hours falling during the subscription period are added
to the delivered quantity of the line.

**Cause**
Commit a6420bdacab3b9c9f440f48db3b442e6b6c7b655 added a new module overwriting
the computation of the delivered quantitiy for recurring lines based on
timesheets, for products with an invoicing policy `delivered_timesheet`.
However, the computation is overriden for all timesheet lines selected by
https://github.com/odoo/enterprise/blob/a6420bdacab3b9c9f440f48db3b442e6b6c7b655/sale_subscription_timesheet/models/sale_order_line.py#L11-L12

A second filtering is applied later
https://github.com/odoo/enterprise/blob/a6420bdacab3b9c9f440f48db3b442e6b6c7b655/sale_subscription_timesheet/models/sale_order_line.py#L21
Which exluces products with an invoicing policy different than
`delivered_timesheet`

**Solution**
For products with a "prepaid" invoicing policy, compute the delivered
quantities the same way: the timesheets delivered during the subscription
period.
Note that as explained in a comment, `Side effect: It won't work for the
first period if the invoice cron never run. (the next invoice date has never
been incremented)`

opw-4509495

closes odoo/enterprise#83383

X-original-commit: 7da3cdd11e2e88265cccea1d6ae43579e34928c9
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>