[FIX] pos_settle_due: never invoice when only settling accounts
Currently, when settling customer's account using the mexican localization,
the session crashes after selecting the payment method.

Steps to reproduce:
-------------------
* Install the mexican localization and switch to one of the company
* Create a new contact of type "Company"
* Create an nvoice for that contact so that he has a due amount set
* Open session
* Select the customer and settle due amount
* Select any payment method
> Observation: Blank screen

Why the fix:
------------
Same issue addressed here but at a different place:
https://github.com/odoo/enterprise/commit/e1e1abdd23cead2bcd7b8170da28161c3f5b2e15

Cause by the fact that when a customer of type "company" is selected we
automatically invoice orders.
https://github.com/odoo/odoo/commit/270779c8250272920d759fbe7b6cfc889e2f5969

Instead of going in the same direction of the first fix, we never set
`to_invoice` to true when we are settling customer accounts. It's pointless
to set invoice to true when settling accounts as it's impossible to
create an invoice without any product. This should also benefit localizations
that invoice by default every order by bypassing the invoice in that case.

opw-4685842

closes odoo/enterprise#83442

X-original-commit: b7f85a3d59c2f2a0df798e7caf7b1f07f4da1e44
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>