[IMP] mrp_workorder: shop floor reintroduce clickable workcenter line
After the removal of the ability to click the workcenter line in [1],
we now re-introduce it with some limitations:

- If the workorder is done, it is marked as disabled and not clickable.
- If the related workcenter is not enabled in the current view
  (i.e. not in the top bar), it is marked as disabled and not clickable.

Also, instead of triggering a filter on the specific MO when clicking,
we now only 'showcase' the selected workorder by showing it as the first
card without filtering out any other records. This 'showcase' is only
persistent until the next time the recordCacheIds are invalidated.

[1] https://github.com/odoo-dev/enterprise/commit/cd9403b838a62717c899fb50f85cd48ca021930f

task-4484941

Part-of: odoo/enterprise#80469
Related: odoo/odoo#200345
Related: odoo/upgrade#7331
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>