[FIX] *: validate quality check for new productions
* quality_mrp, stock_barcode

Steps to reproduce:

- Create a product FP with a bom:
    - 1 x comp
- Create a control check point per product for FP for the manufacturing
  operation type
- Go to the barcode app
- Scan the manufacturing operation type
- Scan FP and set the qty_producing
- Click quality check
> Traceback

Cause of the issue:

By scanning the manufacturing operation in the barcode we create a
"New" manufacturing order that does not have a set id yet. As such, the
`resId` the barcode `MainComponent` is `false`. However, this "id" is
used in the orm call of the `checkQuality`:
https://github.com/odoo/enterprise/blob/79f98e924d2f425535a7ac1919f96ff7ec4a34d8/stock_barcode_quality_control/static/src/components/main.js#L15-L19
As such, we launch a call of the `check_quality`:
https://github.com/odoo/enterprise/blob/79f98e924d2f425535a7ac1919f96ff7ec4a34d8/quality_mrp/models/mrp_production.py#L70-L72
that ends up raising an traceback since the `product_uom` is `False`
on the "unexpected" record `mrp.production(False,)`.

Fix:

Since we just saved the record and hence created the record prior to
this call:
https://github.com/odoo/enterprise/blob/79f98e924d2f425535a7ac1919f96ff7ec4a34d8/stock_barcode_quality_control/static/src/components/main.js#L12-L15
we should rather rely on the new created id.

opw-4644128

closes odoo/enterprise#85024

X-original-commit: 1ef3e274f1ba9582b8f878efb50708bb5b950462
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>