[IMP] mrp_workorder: shop floor show previous admin
When logging out the currently active user in the shop floor, we
remember this last active user and make it more visible in the list by
showing a lock icon and orange left border.

Also, increase the opacity of the cards when there is no currently
logged-in user.

task-4484941

Part-of: odoo/enterprise#80469
Related: odoo/odoo#200345
Related: odoo/upgrade#7331
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>