[FIX] delivery_easypost: Exception for rate_error for multi-package delivery
Before this commit:

Multi-package delivery validation for the carrier 'loomis' results in an error.
Rate is calcuated in the first request for all the packages.
Response of the second request is code 201 but with a warning message of rate already calcuated for the whole shippment.
Which reults in invalid operation error.

After this commit:

Added a check for 'LoomisExpress' to avoid the rate error from blocking validation.
Multi-package delivery is validated correctly.

opw-4650347

closes odoo/enterprise#83118

X-original-commit: b0c4a07f1c220299f52d837c1efcf55a6cdd2be5
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Basit Arshad (baar) <baar@odoo.com>