[IMP] mrp_workorder: shop floor no longer use Layout
As we use no elements of the Layout component in the shop floor, we no
longer use it and structure our own top bar div in mrp_display.xml.
This will facilitate adapting the design of it without having to
override the built-in CSS that ships with the default ControlPanel.

The breadcrumbs component that we still used from the default panel
is now replaced by a custom back button if applicable.

We also add a new splash screen to display when opening the shop floor
for the first time (aka no workcenters set up on this station).

task-4484941

Part-of: odoo/enterprise#80469
Related: odoo/odoo#200345
Related: odoo/upgrade#7331
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>