[FIX] hr_payroll,l10n_hk_hr_payroll: prevent error if work entry type is missing
Currently, an exception is generated when the system tries to find the 'Out of Contract'
work_entry_type after it has been deleted.

Steps to Reproduce:

1 - Install hr_payroll module without demo data
2 - Navigate to hr payroll -> configuration -> work entry types
3 - delete the 'Out of contract' work entry type
4 - Navigate to hr payroll -> Contracts -> Contract
5 - Create a contract with an end date previous to the current.
6 - Change the contract state to 'Running'.
7 - Navigate to hr payroll -> payslips ->  To pay
8 - Try to create Payslips To Pay for created contract
9 - an error occurs

Error:
```ValueError
External ID not found in the system: hr_payroll.hr_work_entry_type_out_of_contract
```
This issue[1] occurs because when the system tries to reference the missing 'Out of Contract'
work entry type it results in a ValueError due to a missing required record. The same issue
[2] occurs in the `l10n_hk_hr_payroll` module.

[1] -
https://github.com/odoo/enterprise/blob/265177dfd281f3aa1d61db6587ddfa37790570fe/hr_payroll/models/hr_payslip.py#L597

[2] -https://github.com/odoo/enterprise/blob/31e56779f8e0011ddc5c1a2821f68c8f041efbe4/l10n_hk_hr_payroll/models/hr_payslip.py#L180

This fix resolves the issue by ensuring that if the `Out of Contract` work entry type is
missing, it returns a false value. A similar fix has been applied in `l10n_hk_hr_payroll`.
Additionally, code has been added to prevent the deletion of any work entry type using
@api.ondelete.

sentry-6277939517

X-original-commit: aa9ee6c
Part-of: odoo/enterprise#82781
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Kishan Kumar Jaiswal (kkuj) <kkuj@odoo.com>