[FIX] l10n_nl_reports_vat_pay_wizard: traceback on pay taxes
Description of the issue this commit addresses:

When paying taxes via the "Pay tax: <period>" button on the miscellaneous
journal of the accounting dashboard, a traceback can be encountered.

Steps to reproduce:

1. Install l10n_nl_reports_vat_pay_wizard.
2. Remove the company vat number.
3. Create and post a closing entry for a period with some tax amount.
4. Go to the accounting dashboard.
5. Click the "Pay tax: <period> in the miscellaneous journal's box.
6. A traceback is raised.

Desired behavior after this commit is merged:

Instead of a traceback, a UserError explains why this is not possible
to the user.

task-4677746

closes odoo/enterprise#82759

X-original-commit: 191769379609fd4e21a0ae325b5efa97a79cc33c
Signed-off-by: John Laterre (jol) <jol@odoo.com>