[IMP] project_enterprise: drag and drop tasks to plan in calendar view
Before this commit, it was not easily possible for the user to plan his
existing tasks inside the calendar view, he was to go another view or
add planned dates in the form view of his tasks.

This commit adds the tasks to plan in desktop mode only to let the user
to drag and drop his tasks unplanned to a calendar cell to directly
plan them.

task-4645814

closes odoo/enterprise#84382

Related: odoo/odoo#207887
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>