[IMP] knowledge: allow partner to remove same level member
This commit allows the user to remove its member from an article which has the
same permission level. Previously, when a member had the same permission level
as the article it was on couldn't be removed by the user linked to the member.

The issue is that if a user joined an article just to be able to see it now it
couldn't remove its member from it. Even though there wasn't any rights
escalation being done, e.g. a reader member on a readonly article is
functionnally the same as having no member.

task-4273102

Part-of: odoo/enterprise#76261
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>