[REF] iot: remove duplicate connected field
In #81180, a `connected_status` field was added to the `iot.device`
model in order to show the status as a banner widget. However, this
meant that there were two separate fields for the connection status,
`connected` and `connected_status`.

This commit removes the `connected` field and replaces all usages of it
with `connected_status`.

task-4652455

closes odoo/enterprise#82720

Related: odoo/upgrade#7484
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>