[IMP] *: replace fa-ellipsis-* occurrences with `oi-ellipsis` icons
*: account_reports, appointment, knowledge, l10n_us_reports, planning,
sign, social, social_crm, spreadsheet_edition, web_studio

This commit replaces all the existing occurrences of `fa-ellipsis-* with the
new `oi-ellipsis-*` icons.

Prior to this commit, we were relying on Font Awesome ellipsis icons which
comes with squared dots that look quite bulky.
As the ellipsis icon is shown in some key places, we decided to slightly
tweak it to improve the visual result.

Note that the fa-ellipsis-* still exist and remain available but in
order to maintain the consistency, we highly recommend to use
`oi-ellipsis-v` and `oi-ellipsis-h` from now on.

task-4551038

closes odoo/enterprise#82735

Related: odoo/odoo#204473
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>