[IMP] web_studio: edit_view uses KeyedXMLDiffer
Before this commit there were two ways to make inheriting views more compact in terms of xpath:
- The KeyedXMLDiffer (commit)[odoo/enteprise@075f7d35c0ac5722365d222a184d6dc9b8c64677], used until now only in the case of pure QWEB archs (reports)
- the `def normalize´ function in ir_ui_view

The second option was cool but insufficient:
- The algorithm was very complicated and relied on textual diff.
Identification of nodes was "best effort"
- There were bugs in how the target of a change was computed, and how changes were grouped together

This commit aims at using the KeyedXMLDiffer for both use cases.
The code should be simpler -- it is never simple to browse trees at all, not to mention the limitations of Odoo's inheritance semantics (the xpath node)
A quick performance check yielded an improvement of almost an order of magnitude (6.5 times on small rather small trees)
The test suite is self-contained and clearer. It should be easier to debug, fix and test altogether.

Note that thanks to (commit)[odoo/odoo@040aec546d29a511d7f117b676796bd2b1cad285], it is now possible to get the result of an inheritance tree
without writing onto the view to exclude from it.
A quick performance check shows that this certainly has an effect, but is not the source of the whole gain, rather, the change in algorithm is.
Bugs that involve misplacing an element are now solved

Canonical normalized views see some changes though.
Because we introduce the concept of subtree (a subview within a field), xpaths targetting identifiable nodes are
of the form: `/[subtree]//tag[identifier]`. For fields it translates to: `/form//field[name="field_name"]`

task-4853076

closes odoo/enterprise#82729

Signed-off-by: Hubert Van De Walle <huvw@odoo.com>