[FIX] sale_amazon: reset offer information when discrepancy
When the fulfillment channel changes from Amazon, the user currently
must delete the offer to create a new one, as they can't change the
amazon_feed_ref.

(Please note that due to an API change, this field does in fact include
a JSON containing the product type, fulfillment channel, and the feed
reference.)

Indeed, when the listing is changed, Amazon doesn't get the information.
This means we will continue to update the inventory if that offer was in
FBM. Or not updating the inventory of a now FBM offer.

While that second case would only require the user to delete and
recreate a new offer (which would be a bother but not that problematic),
sending a quantity to a FBA offer would trigger a well known problem
of ghost listing on Amazon, where Amazon will consider having both
channels, and randomly set an order as either FBA or FBM, until that
ghost listing quantity drop down to 0.

When catching a discrepancy, we'll thus reset the value of the
amazon_feed_ref, so that it'd be pulled again later if needed by the sync
inventory with, this time, the correct value. This would also stop the
inventory synchronization of a potentially FBA offer.

As for any stock synchronization that might happen between the listing
changes from Amazon side and the first order received in FBA, we decided
against forcing the FBM quantity to 0 on all FBA offer every time, to
avoid sending too many useless calls for a corner case. As the product
was until now sold in FBM, in most cases, it can still be until the end
of the ghost listing. If not, our support is aware of it and can force the
quantity to 0 again.

opw-4681035

closes odoo/enterprise#82786

X-original-commit: 6b5acdde9998fb1ee0d4d4aee0ab2375e50d7703
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>